Add disabled notice to AnalyzeChangeJob
We've disabled `AnalyzeChangeJob` via configuration in production. Since we also expect to never make use of it again, I've added a note in the code (I opted not to delete it in case it needs to get turned on in the future -- someone could always go get it back from the git history, but since there is no technical maintainer to replace me, I'm worried that's too big of a barrier.)

Fixes #1090.