Use effective_status for Page#status
Pages have a `status` attribute that is meant to be a convenient way to figure out what the current status of a page is, skipping over short-lived, erroneous errors from a single snapshot. A while back, we added "effective" status codes to versions, which attempt to figure out if a version with a 200 status code actually should have been an error code (surprisingly common!). We should have updated Page's status calculation to use that, since the page is really about what is effectively the current state, but missed it at the time.