Fix regression caused by change to `find_or_create_by`
We used to rely on this method raising `RecordNotUnique`, but now it raises `RecordNotFound` in this case. See: https://github.com/rails/rails/issues/51149

It turns out our usage was kind of not the intended usage (you should only be passing in the fields that are meant to be unique). We could try to rejigger this, but it would be unnecessarily more complicated. The structure of this model and this usage just doesn't really match what Rails maintainers are thinking of for this method.