Use setup/teardown *blocks*, not methods
Something changed subtly that broke places where we used `def setup` in tests, and now it appears that we always need to use the block form (`setup do ...`). I couldn't find much information on this, although it seems like the general recommendation is to use the blocks when inheriting from `ActiveSupport::TestCase`. It's possible the method form was never really supported, and only happened to be working for us before.