Fix new type error in `where_in_interval` query
The previous approach started raising `PG::InvalidTextRepresentation` errors because we were using a float to query an integer column. This looks pretty similar to https://github.com/rails/rails/issues/51617, which is apparently caused by https://github.com/rails/rails/pull/51139.

At any rate, an easy solution here is to switch to using Arel. (Another is to check the column type and call `to_i` on the value if the column we are working with is an integer.)