Link `data` directory in Sprockets manifest
The last release seems to be running into some issues with Sprockets failing to recognize the `swagger.yml` file (even though it successfully precompiles). In Sprockets 4 there were changes to how config options are handled vs. what's in the `manifest.js` file, so I'm adding it to `manifest.js` to see if that solves it.