Fix typo that is killing production
Apparently Sprockets used to transparently handle multiple file extensions for a given file type, so it made an old typo where we used to reference a `.yaml` file as `.yml` just work. In a recent update, that seems to have stopped!

I originally thought the problem here was in the manifest (see #1162), but it turns out it was actually just this typo.