Add `network_error` field to Version model (#1184)
This gives us a way to model the server itself being down instead of responding with an HTTP 4xx or 5xx error. In the very early times (when this DB was focused around data from Versionista and Pagefreezer), we shoved errors and HTTP statuses in `source_metadata` because we often didn't have a clear representation of different kinds of errors. We later switched to 100% Internet Archive as a source, and didn't consider the fact that they don't track network-level errors since we were no longer seeing whole hostnames disappear at that point. Now at the start of 2025, we're seeing a lot of servers simply being shut down, and need to be able to clearly represent that.

For captures of network-level failures (i.e. no HTTP response message was received), we can record a version with:
- `body_url = nil`
- `body_hash = nil`
- `status = nil`
- `network_error = '<freeform text representation of the error>'`

This also adds a nice template for representing the raw content of a network error so analysts can see something meaningful and so that existing diff and display machinery doesn't break.