Ensure network errors get counted for page status
In #1184, I added the ability to track versions that represent network errors instead of valid HTTP reponses. I forgot about calculating the page's effective status, though, where these versions were getting ignored (since they have nil status). Ideally, I think we'd set a status of `0`, but there are some validations that disallow that and other places I need to be careful of causing problems. So the easiest thing for now is to treat a network error as a 404. Close enough to do the job.