Fix import format upgrade logic
Several years back, we changed the version import logic to matche the changing (and more HTTP response-like) fields for `Version`. When doing so, we attempted to be backwards compatible with the old import format. It turns out that didn't quite work with the `page_url` field, though! The cleanup/upgrade came too late in the process, and a different part of the code was still requiring `page_url` when it should have required `url`. This fixes the issue.