Build Docker images as part of normal CI (#1199)
We previously only built the Docker images for this project as part of the release process (because building and publishing were integrated). That recently led to an issue where a subtle change to our base image between versions caused a release to fail. This ensures we build the Docker images whenever we run tests so we get a heads-up on issues like this.

This is a follow on to #1195.