Fix incorrect diff_types in swagger docs (#1205)
At some point, we changed the names of some of the diffs and also added more in web-monitoring-diff. Since the diff endpoint in this project is mostly just a pass-through to that, those changes made the docs here inaccurate. This brings the docs for the diff endpoint up-to-date with reality.

Fixes #1204.

Co-authored-by: Rob Brackett <rob@robbrackett.com>