Do a better job normalizing URLs
We've had a bunch small issues lately with multiple pages and other oddball matching issues because the way we normalize URLs is not expansive enough (or totally consistent between `Page` and `PageUrl`!). This takes a more serious crack at normalization and also switches to Rails's new(ish) normalization tooling so we can clean up a lot of callsites because normalization now Just Happens(tm) at the right times.

Ideally we could lean on Addressable or Surt for real robust normalization here, but Addressable requires more careful investigation (it would alter over 1,000 of our URLs in ways that are probably correct but I am 100% confident about) and Surt turns out to need some more work in order to be safe (see #1215 for more). So for now we have our own custom normalization.