Lots of SURT cleanup, including empty paths (#1215)
This cleans up and tweaks a whole bunch of little things in our SURT implementation, including making sure empty paths get transformed into a single slash ("/") for #1209.

This was originally just meant to address #1209, but I kept running into little issues that have sat unfixed for surprisingly long. 😬 As such, I’m keeping normalization of `Page#url` and `PageUrl#url` in a separate PR.

This also adds a concept of “safe” options, which was originally meant for the URL normalization above. My thinking was that this would only apply the options that are “safe” and which are guaranteed not to change the meaning of the URL. However, it turns out we *always* do some unsafe things that have no option at all! So “safe” canonicalization is not ready yet (although I’d like to get there). There’s just not enough time today to really go over all this with a fine-toothed comb, so I added a bunch of notes inline about remaining issues that were obvious to me.