Do a better job normalizing URLs (#1216)
We've had a bunch small issues lately with multiple pages and other oddball matching issues because the way we normalize URLs is not expansive enough and not totally consistent between `Page` and `PageUrl`(!). This takes a more serious crack at normalization and also switches to Rails's new(ish) normalization tooling so we can clean up a lot of call-sites because normalization now Just Happens™ at the right times.

Ideally we could lean on Addressable or Surt for a real robust normalization algorithm, but Addressable requires more careful investigation (it would alter over 1,000 of our URLs in ways that are probably correct but I am 100% confident about) and Surt turns out to need some more work in order to be safe (see #1215 for more). So for now we have our own custom normalization.

This also includes a data migration to update all our `PageUrl` instances and clean up any duplication issues this might cause.

This is second part of #1209, and fixes it in conjunction with #1215.