Try not to cache diff errors in CDN (#1219)
We used to use `stale?` to avoid re-computing diffs when possible, but if we raised an error afterward (during diffing), we would wind up with some fairly extreme caching headers getting sent out with the error! `stale?` sets a lot of different caching mechanisms when it's called and doesn't give you an easy way to roll them back. This solves the issue by creating a `when_stale()` method that takes a block. It keeps track of the original header values and resets them if the block raises an error instead of returning a useful value, that way an error handling code can safely set (or not set) cache without worrying about what other cache features have already been messed with.

This also moves some of the more complicated cache-control header option setting into the `stale?` call so it gets rolled back in the same way (it was previous set separately in an earlier step).