Drop support for `?different` query param
This covers part 1 of #1208 and generally removes support for the `different` query param, which doesn't behave like you would probably expect and regularly causes problems. It was designed in a world where we expected to have such frequent and continuous captures of pages that any time we had a "different" version, we knew it had *just* changed. The reality is that you can never realistically get captures frequent enough to do that even in practical terms, and in theoretical terms it's just wrong. We've long since switched to always querying in a way that removes checking the `different` field, and it's time to just remove it from the API.

This leaves it in place in ONE spot of the API, which does not return an actual list of versions and thus does not have the same problems: `GET /pages?capture_time=abc...def` implicitly checks the `different` column because the goal is to tell you what pages had observed changes in that time period. We could back off that idea and just have it check whether there were any captures, but I think the valuable use cases for that are pretty rare (especially if we are mainly focused on pages where `active=true`, and we expect that every page should have captures at roughly the same times).