Update Swagger docs for `Version` and `Page` (#1235)
I forgot to update these when I added `network_error`, but it looks like there have been a *lot* of updates and changes that have not hit these docs. This updates the fields in the `Page` and `Version` models, although there is still more that's not quite up-to-date here, I think.