Fix warnings about deprecated `:unprocessable_entity` (#1288)
The IETF revised the name of status code 422 to "Unprocessable Content" and Rack also renamed the symbol used for the same status code. The previous symbol (`:unprocessable_entity`) has been deprecated and now logs lots of warnings. This fixes all the uses of it in our code, although we still get warnings from within Rails and WebMock.