Remove deprecated import values from tests, docs (#1291)
We *long* ago deprecated and renamed some fields in versions and from the versions import code, but still had some lingering references in the docs, tests, and seed data. This removes them from those places (they are still supported but deprecated in the import code). Ultimately, I don't think we've *used* these (except in the seed data) for years, so I'd also like to stop supporting them as deprecated in the importer, but that's needs to be a separate, later step.