Support getting pages with empty `?url` param
If we try to filter pages by URL but the filter is the empty string (i.e. the request was for `/pages?url=`), we should treat that as not filtering by URL at all. I think this is what I'd originally intended, and messed up because empty strings are truthy in Ruby. Similarly, querying for all pages with `?url=*` works but is kinda dumb; we should just not add any URL-related conditions at all to improve performance in that case.

This fixes some issues I ran into with multiple other tools today, trying to provide simplified options for things and mishandling "no URL filtering specified" by passing an empty string. Oops.