Disable Stale bot
Long ago, we added the stale issue bot to help manage the influx of a lot of ideas and brainstorming against a small implementation team. Today, the web-monitoring tools mostly just have one maintainer (me) and are not flooded with a lot of issues and ideas. Long-lived issues are more common, and they are helpful for me to keep track of things over the slow pace and timeframe in which I can individually address stuff. Stale bot feels like it is getting more in the way than helping in the current moment, so I’m disabling it.